+++ src/actions/index.js
@@ -1,9 +1,0 @@
-import { ActionTypes } from '../constants';
-
-export function validate() {
-  return { type: ActionTypes.FETCH_VALIDATE };
-}
-
-export function invalidate() {
-  return { type: ActionTypes.FETCH_INVALIDATE };
-}