+++ src/components/fetch.js
@@ -1,15 +1,17 @@
+/* eslint-disable no-underscore-dangle */
+
 import React, { Component, PropTypes } from 'react';
 import invariant from 'invariant';
 import hoistStatics from 'hoist-non-react-statics';
-import { invalidate } from '../actions';
 import fetchData from '../utilities/fetch-data';
 import getDisplayName from '../utilities/get-display-name';
 import Spinner from './spinner';
+import Glitch from './glitch';
 
-// Global settings, overwrite using `fetch.setup` function.
-const fetchSettings = {
-  activityIndicator: Spinner,
-  selectFetchingState: state => state.fetching,
+const globalSettings = {
+  forceInitialFetch: false,
+  renderFailure: (/* error */) => <Glitch />,
+  renderLoading: () => <Spinner />,
 };
 
 const storePropTypes = PropTypes.shape({
@@ -18,12 +20,10 @@ const storePropTypes = PropTypes.shape({
   getState: PropTypes.func.isRequired,
 });
 
-export default function fetch(getInitialAsyncState) {
-  return function wrapWithFetch(
-    RouteComponent,
-    ActivityIndicator = fetchSettings.activityIndicator
-  ) {
-    const displayName = `Fetch(${getDisplayName(RouteComponent)})`;
+export default function fetch(getInitialAsyncState, options = {}) {
+  return function wrapWithFetch(WrappedComponent) {
+    const displayName = `Fetch(${getDisplayName(WrappedComponent)})`;
+    const finalSettings = Object.assign({}, globalSettings, options);
 
     class Fetch extends Component {
       static displayName = displayName;
@@ -48,51 +48,45 @@ export default function fetch(getInitialAsyncState) {
           'Either wrap the root component in a `<Provider>`, ' +
           `or explicitly pass "store" as a prop to "${displayName}".`
         );
-
-        invariant(fetchSettings.selectFetchingState(store.getState()),
-          'Expected the fetching state to be available as `state.fetching`. ' +
-          'or as the custom expression you can specify as `selectFetchingState` ' +
-          'with the `fetch.setup()` interface. ' +
-          'Ensure you have added the `fetchReducer` to your store\'s ' +
-          'reducers via `combineReducers` or whatever method you use to isolate ' +
-          'your reducers.'
-        );
       }
 
       state = {
+        error: null,
+        hasError: false,
         isFetching: false,
       };
 
       componentWillMount() {
-        const state = fetchSettings.selectFetchingState(this.store.getState());
+        if (this.store.__didInitialLoad__ || finalSettings.forceInitialFetch) {
+          this.setState({ isFetching: true, hasError: false, error: null });
 
-        // Guard to ensure we do not attempt to reload route state when data
-        // is hydrated on the client-side from the serve-side.
-        if (!state.shouldFetch) {
-          // Invalidate guard so that subsequent route components fetch data
-          // from the server-side.
-          this.store.dispatch(invalidate());
-        } else {
-          this.setState({ isFetching: true });
           fetchData(this.store, getInitialAsyncState).then(() => {
             this.setState({ isFetching: false });
+          }).catch((error) => {
+            this.setState({ isFetching: false, hasError: true, error });
           });
         }
+
+        this.store.__didInitialLoad__ = true;
       }
 
       render() {
-        if (this.state.isFetching && ActivityIndicator) {
-          return <ActivityIndicator {...this.props} />;
+        if (this.state.isFetching) {
+          return finalSettings.renderLoading();
         }
 
-        return <RouteComponent {...this.props} />;
+        if (this.state.hasError) {
+          return finalSettings.renderFailure(this.state.error);
+        }
+
+        return <WrappedComponent {...this.props} />;
       }
     }
 
-    return hoistStatics(Fetch, RouteComponent);
+    return hoistStatics(Fetch, WrappedComponent);
   };
 }
 
 fetch.setup = (options) => {
-  Object.assign(fetchSettings, options);
+  Object.assign(globalSettings, options);
 };