+++ src/constants/index.js
@@ -1,4 +1,0 @@
-export const ActionTypes = {
-  FETCH_VALIDATE: '@@fetch/VALIDATE',
-  FETCH_INVALIDATE: '@@fetch/INVALIDATE',
-};