+++ src/index.js
@@ -1,3 +1,4 @@
 export { default as fetch } from './components/fetch';
+export { default as Spinner } from './components/spinner';
+export { default as Glitch } from './components/glitch';
 export { default as fetchRouteData } from './utilities/fetch-route-data';
-export { default as fetchReducer } from './reducers';