+++ test/specs/components/fetch.js
@@ -1,101 +1,140 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import createMockStore from '../../utilities/create-mock-store';
-import { invalidate } from '../../../src/actions';
+import waitFor from '../../utilities/wait-for';
 import fetch from '../../../src/components/fetch';
 import Spinner from '../../../src/components/spinner';
+import Glitch from '../../../src/components/glitch';
 
-const RouteComponent = () => <div className="route" />;
+const Component = () => <div className="route" />;
 
-function createWrappedComponent(getInitialAsyncState = sinon.stub().returns(Promise.resolve())) {
-  const WrappedComponent = fetch(getInitialAsyncState)(RouteComponent);
-  return { getInitialAsyncState, WrappedComponent };
+function mountWithStore(node) {
+  const store = createMockStore({ stateful: true });
+  const wrapper = mount(node, { context: { store } });
+  return { store, wrapper };
 }
 
-function mountWithStore(node, store) {
-  return mount(node, { context: { store } });
-}
-
-describe('fetch(getInitialAsyncState)(RouteComponent[, ActivityIndicator])', () => {
-  it('should render its default activity indicator while fetching', () => {
-    const store = createMockStore({ fetching: { shouldFetch: true } });
+describe('fetch(getInitialAsyncState, options)', () => {
+  it('should render its default loading component while fetching', () => {
     const getInitialAsyncState = sinon.stub().returns(new Promise(() => {}));
-    const WrappedComponent = fetch(getInitialAsyncState)(RouteComponent);
-    const wrapper = mountWithStore(<WrappedComponent />, store);
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    const { wrapper } = mountWithStore(<WrappedComponent />);
     expect(wrapper.find(Spinner)).to.have.length(1);
   });
 
-  it('should render its loading component while fetching', () => {
+  it('should render a custom loading component while fetching', () => {
     const ActivityIndicator = () => <div className="loading" />;
-    const store = createMockStore({ fetching: { shouldFetch: true } });
     const getInitialAsyncState = sinon.stub().returns(new Promise(() => {}));
-    const WrappedComponent = fetch(getInitialAsyncState)(RouteComponent, ActivityIndicator);
-    const wrapper = mountWithStore(<WrappedComponent />, store);
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+      renderLoading: () => <ActivityIndicator />,
+    })(Component);
+    const { wrapper } = mountWithStore(<WrappedComponent />);
     expect(wrapper.find(ActivityIndicator)).to.have.length(1);
+    expect(wrapper.find(Spinner)).to.have.length(0);
   });
 
-  it('should render its route component after fetching', (done) => {
-    const store = createMockStore({ fetching: { shouldFetch: true } });
-    const { WrappedComponent } = createWrappedComponent();
-    const wrapper = mountWithStore(<WrappedComponent />, store);
-    // defer to next tick to allow setState(...) to be picked up by event loop
-    setTimeout(() => {
-      expect(wrapper.find(RouteComponent)).to.have.length(1);
-      done();
-    }, 0);
+  it('should render its own failure component when an unhandled error occurs', (done) => {
+    const getInitialAsyncState = sinon.stub().returns(Promise.reject());
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    const { wrapper } = mountWithStore(<WrappedComponent />);
+    waitFor(
+      () => wrapper.state('isFetching') === false,
+      'Data fetching timed out',
+      () => {
+        expect(wrapper.find(Glitch)).to.have.length(1);
+        expect(wrapper.find(Component)).to.have.length(0);
+        done();
+      }
+    );
   });
 
-  it('should invalidate fetch state when rendered', () => {
-    const store = createMockStore({ fetching: { shouldFetch: false } });
-    const { WrappedComponent } = createWrappedComponent();
-    mountWithStore(<WrappedComponent />, store);
-    expect(store.dispatch).to.have.been.calledOnce;
-    expect(store.dispatch).to.have.been.calledWithExactly(invalidate());
+  it('should render route component after data fetching is successful', (done) => {
+    const getInitialAsyncState = sinon.stub().returns(Promise.resolve());
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    const { wrapper } = mountWithStore(<WrappedComponent />);
+    waitFor(
+      () => wrapper.state('isFetching') === false,
+      'Data fetching timed out',
+      () => {
+        expect(wrapper.find(Component)).to.have.length(1);
+        done();
+      }
+    );
   });
 
-  it('should not fetch data requirements when `{ shouldFetch: false }`', () => {
-    const store = createMockStore({ fetching: { shouldFetch: false } });
-    const { getInitialAsyncState, WrappedComponent } = createWrappedComponent();
-    mountWithStore(<WrappedComponent />, store);
+  it('should not fetch data requirements when `{ forceInitialFetch: false }`', () => {
+    const getInitialAsyncState = sinon.stub();
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: false,
+    })(Component);
+    mountWithStore(<WrappedComponent />);
     expect(getInitialAsyncState).to.not.have.been.called;
   });
 
-  it('should fetch data before rendering when `{ shouldFetch: true }`', () => {
-    const store = createMockStore({ fetching: { shouldFetch: true } });
-    const { getInitialAsyncState, WrappedComponent } = createWrappedComponent();
-    mountWithStore(<WrappedComponent />, store);
+  it('should fetch data before rendering when `{ forceInitialFetch: true }`', () => {
+    const getInitialAsyncState = sinon.stub().returns(Promise.resolve());
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    mountWithStore(<WrappedComponent />);
     expect(getInitialAsyncState).to.have.been.calledOnce;
   });
 
-  it('should throw when store is configured incorrectly', () => {
-    const store = createMockStore();
-    const { WrappedComponent } = createWrappedComponent();
-    expect(mount.bind(null, <WrappedComponent />)).to.throw;
-    expect(mountWithStore.bind(null, WrappedComponent, store)).to.throw;
+  it('should pass store dispatch and current state to getInitialAsyncState', () => {
+    const getInitialAsyncState = sinon.stub().returns(Promise.resolve());
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    const { store } = mountWithStore(<WrappedComponent />);
+    expect(getInitialAsyncState).to.have.been.calledOnce;
+    expect(getInitialAsyncState).to.have.been.calledWithExactly(store.dispatch, store.getState());
   });
 
-  it('should not throw an error when store is configured correctly', () => {
-    const store = createMockStore({ fetching: { shouldFetch: false } });
-    const { WrappedComponent } = createWrappedComponent();
-    expect(mountWithStore.bind(null, WrappedComponent, store)).to.not.throw;
+  it('should throw when store is not in scope', () => {
+    const getInitialAsyncState = sinon.stub();
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: false,
+    })(Component);
+    expect(mount.bind(null, <WrappedComponent />)).to.throw;
   });
 });
 
 describe('fetch.setup(options)', () => {
-  it('should overwrite default activity indicator', () => {
-    const MyActivityIndicator = () => <div />;
-    fetch.setup({ activityIndicator: MyActivityIndicator });
-    const store = createMockStore({ fetching: { shouldFetch: true } });
+  it('should overwrite default loading component', () => {
+    const ActivityIndicator = () => <div className="loading" />;
+    fetch.setup({ renderLoading: () => <ActivityIndicator /> });
     const getInitialAsyncState = sinon.stub().returns(new Promise(() => {}));
-    const { WrappedComponent } = createWrappedComponent(getInitialAsyncState);
-    const wrapper = mountWithStore(<WrappedComponent />, store);
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    const { wrapper } = mountWithStore(<WrappedComponent />);
     expect(wrapper.find(Spinner)).to.have.length(0);
-    expect(wrapper.find(MyActivityIndicator)).to.have.length(1);
+    expect(wrapper.find(ActivityIndicator)).to.have.length(1);
   });
 
-  it('should overwrite default selector for fetching state', () => {
-    const store = createMockStore({ custom: { shouldFetch: false } });
-    const { WrappedComponent } = createWrappedComponent();
-    expect(mountWithStore.bind(null, WrappedComponent, store)).to.not.throw;
+  it('should overwrite default failure component', (done) => {
+    const Boom = () => <div className="boom" />;
+    fetch.setup({ renderFailure: () => <Boom /> });
+    const getInitialAsyncState = sinon.stub().returns(Promise.reject());
+    const WrappedComponent = fetch(getInitialAsyncState, {
+      forceInitialFetch: true,
+    })(Component);
+    const { wrapper } = mountWithStore(<WrappedComponent />);
+    waitFor(
+      () => wrapper.state('isFetching') === false,
+      'Data fetching timed out',
+      () => {
+        expect(wrapper.find(Glitch)).to.have.length(0);
+        expect(wrapper.find(Boom)).to.have.length(1);
+        done();
+      }
+    );
   });
 });